v0.14.0
New Editor 🚀
We've ditched the card editor modal and replaced it with an inline editing mode
This mode can still be accessed with double clicking the card or pressing the edit button in the card bar.
New card template
Instead of a 'Untitled card' with placeholder text, when you add a card to your MDyna boards, the default text is a shortcuts guide.
In the future we intend to use this to introduce more card templates 😉
New color picker
The MDyna card color picker has a new UI and functionalities
You can now choose HEX colors for your cards
New label picker
Also with a new UI and with a new functionality
Users are able to click on labels to remove them from the current card
Clicking on A will create a new card on the current board
The hover animation on the cards is now gone. That visual aid is now in place for when a card is being edited

Bugs 🐞
Fix emoji plugin for when there are no emojis matching the string
'ESC' shortcut was not working when focusing cards. That is now fixed.
Fix generic errors in gistsService and board picker component.
Fix issue which was causing the tooltips to flicker in certain positions.
Fix settings UI overflow.